Swap embedded-kafka for testcontainers
